First round of updates/refactoring; produced a JSON representation of the PEAP data for use with a basic API.
(1) Updated read_url_list to fetch the 'url list' from the actual current PEAP list
instead of a curated local file; thus, the input parameter is the now the URL of
the remote PEAP list instead of a path to a local file contains a list of URLs.
As a result of this change, there is no longer a strictly offline mode of operation.

(2) Updated process_url to *always* use the locally cached copy of the page unless it
doesn't exist.  This obviates the need for use_local.  Considering an additional parameter
to force the local cache to update, thus effectively reversing the semantics of use_local.

(3) Added JSON output in addition to CSV.  The JSON output excludes fields such as 'business_information_raw'
and 'business_information_addl', and also nicely formats field names for use in JSON.
The JSON object generated by this process serves as the basis for peora-il-equity-data-api.

(4) Updated the Jupityr notebook IAW the above.

(5) Regenerated data.

(5) Whitespace/formatting changes throughout.